Working with EDF again...

+++ src/EDF/record.h
@@ -1,10 +1,12 @@
 #ifndef RECORD_H
 #define RECORD_H 0
+#include <vector>
+#include <map>
 
 typedef struct
 {
 	std::map<std::string, std::string> header;
-	std::vector<float> records;
+	std::vector<int> records;
 	size_t number_samples;
 } RECORD;
